requirements.txt
PyMuPDF
Pillow
python-dotenv
google-generativeai
pandas